requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Templates
jinja2==3.1.2

# Data Validation
pydantic==2.5.0
email-validator==2.1.0

# YAML Processing
pyyaml==6.0.1

# Security & Crypto
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
itsdangerous==2.1.2

# HTTP Client (for health checks)
httpx==0.25.1

# Testing Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0

# E2E Testing (Playwright)
playwright==1.40.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
